move IP generators into an abstract base class
